% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_2rm.R
\name{agreguj_cawi_ucz_2rm}
\alias{agreguj_cawi_ucz_2rm}
\title{Obliczanie wskaznikow z 2 rundy monitoringu na poziomie zagregowanym}
\usage{
agreguj_cawi_ucz_2rm(wskazniki, grupy)
}
\arguments{
\item{wskazniki}{ramka danych z wynikami 2 rundy monitoringu}

\item{grupy}{ramka danych zawierająca definicje podziałów na grupy -
np. zwrócona przez funkcję \code{\link{utworz_grupowanie_ze_zmiennej}}}
}
\value{
data frame
}
\description{
Funkcja oblicza wartości wskaźników na poziomie zagregowanym
na podstawie ramki danych z wynikami ankiety CAWI.
}
\seealso{
\code{\link{agreguj_wskazniki}} oraz przekazywane do niej funkcje
używane do obliczania konkretnych wskaźników zagregowanych:
\itemize{
  \item{\code{\link{liczba_zbadanych}},}
  \item{\code{\link{dane_szkoly}},}
  \item{\code{\link{liczba_kobiet}},}
  \item{\code{\link{firma_badawcza}},}
  \item{\code{\link{formy}},}
  \item{\code{\link{zawod_liczebnosc}},}
  \item{\code{\link{szkola_1_wyboru}},}
  \item{\code{\link{ponowny_wybor}},}
  \item{\code{\link{przygotowanie_do_zawodu}},}
  \item{\code{\link{przyg_zawodu_prakt_PL}},}
  \item{\code{\link{przyg_zawodu_prakt_niePL}},}
  \item{\code{\link{przyg_zawodu_zaj_PL}},}
  \item{\code{\link{przyg_zawodu_zaj_niePL}},}
  \item{\code{\link{przyg_zawodu_zaj_szkola}},}
  \item{\code{\link{przyg_zawodu_zaj_ckp}},}
  \item{\code{\link{nauka_zawod}},}
  \item{\code{\link{plany_6m}},}
  \item{\code{\link{czy_plany_eduk}},}
  \item{\code{\link{plany_eduk_tak}},}
  \item{\code{\link{plany_eduk_nie}},}
  \item{\code{\link{praca_zarobkowa}},}
  \item{\code{\link{praca_poza_wyuczonym}},}
  \item{\code{\link{brak_pracy}},}
  \item{\code{\link{mlodociani_praca}},}
}
}
