% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_1rm.R
\name{agreguj_wskazniki_1rm_szk}
\alias{agreguj_wskazniki_1rm_szk}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
agreguj_wskazniki_1rm_szk(
  wskazniki,
  progAnonimizacji = 10,
  wykluczGrupeZGrupyOdniesienia = FALSE
)
}
\arguments{
\item{wskazniki}{ramka danych ze wskaźnikami na poziomie idywidualnym
zwracana przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}}

\item{progAnonimizacji}{liczba - wskazniki obliczone na podstawie
mniejszej liczby badanych zostaną zamienione na braki danych
(p. \code{\link{anonimizuj_wskazniki}})}

\item{wykluczGrupeZGrupyOdniesienia}{wartość logiczna - czy obserwacje
z analizowanej grupy powinny zostać wykluczone z grupy odniesienia}
}
\value{
lista dwóch ramek danych:
\itemize{
  \item{\code{grupy} - ramka danych zawierająca wskaźniki obliczone dla
        poszczególnych grup,}
  \item{\code{grupyOdniesienia} - ramka danych zawierająca wskaźniki
        obliczone dla odpowiadających im grup odniesienia.}
}
}
\description{
Funkcja oblicza wartości wskaźników na poziomie szkoły
na podstawie ramki danych ze wskaźnikami z poziomu indywidualnego
(zwrócnej przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}).
}
\seealso{
\code{\link{agreguj_wskazniki_1rm}}
i code{\link{utworz_grupowanie_ze_zmiennej}}, które wykonują większość pracy
}
