% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{agreguj_wskazniki_szk_branza}
\alias{agreguj_wskazniki_szk_branza}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
agreguj_wskazniki_szk_branza(wskazniki, progAnonimizacji = 10)
}
\arguments{
\item{wskazniki}{ramka danych ze wskaźnikami na poziomie idywidualnym
zwracana przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}}

\item{progAnonimizacji}{liczba - wskazniki obliczone na podstawie
mniejszej liczby badanych zostaną zamienione na braki danych
(p. \code{\link{anonimizuj_wskazniki}})}
}
\value{
data frame
}
\description{
Funkcja oblicza wartości wskaźników na poziomie branży
w ramach szkoły na podstawie ramki danych ze wskaźnikami z poziomu
indywidualnego (zwrócnej przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}).
}
\details{
Funkcja pozwala na łatwe obliczenie wskaźników zagregowanych na
poziomie branży w ramach szkoły w sposób kompatybilny z zestawieniem
wskaźników w grupach porównawczych zdefiniowanych przez kombinację branży
i typu szkoły, które tworzy funkcja \code{\link{agreguj_wskazniki_typ_szk_branza}}.
}
\seealso{
\code{\link{agreguj_wskazniki_typ_szk_branza}}
oraz \code{\link{agreguj_wskazniki_1rm}}, która wykonuje większość pracy
}
