% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{agreguj_wskazniki_typ_szk_branza}
\alias{agreguj_wskazniki_typ_szk_branza}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
agreguj_wskazniki_typ_szk_branza(wskazniki)
}
\arguments{
\item{wskazniki}{ramka danych ze wskaźnikami na poziomie idywidualnym
zwracana przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}}
}
\value{
data frame
}
\description{
Funkcja oblicza wartości wskaźników na poziomie branży w ramach
typu szkoły na podstawie ramki danych ze wskaźnikami z poziomu indywidualnego
(zwrócnej przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}).
}
\details{
Funkcja pozwala na łatwe obliczenie wskaźników zagregowanych na
poziomie branży w ramach typu szkoły w sposób kompatybilny z zestawieniem
wskaźników na poziomie branż w ramach szkół, które tworzy funkcja
\code{\link{agreguj_wskazniki_szk_branza}}.
}
\seealso{
\code{\link{agreguj_wskazniki_szk_branza}}
oraz \code{\link{agreguj_wskazniki_1rm}}, która wykonuje większość pracy
}
