% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonimizuj_wskazniki.R
\name{anonimizuj_wskazniki}
\alias{anonimizuj_wskazniki}
\title{Anonimizacja wskaznikow na poziomie zagregowanym}
\usage{
anonimizuj_wskazniki(wskazniki, prog = 10)
}
\arguments{
\item{wskazniki}{ramka danych ze wskaźnikami na poziomie zagregowanym
zwracana przez funkcję \code{\link{agreguj_wskazniki_1rm}}}

\item{prog}{liczba - minimalna liczba badanych niewymagająca anonimizacji}
}
\value{
data frame
}
\description{
Funkcja anonimizuje zestawienie wskaźników na poziomie
zagregowanym poprzez przypisanie wartości \code{NA} wszystkim elementom
każdego wskaźnika w danej grupie, dla którego wartość elementu \code{n} jest
mniejsza niż \code{prog}. Kolumny nie będące listami (oraz kolumny-listy,
których elementy nie są listami zawierającymi element o nazwie \code{n})
pozostaną w zwróconej ramce danych niezmienione.
}
