% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{bezrobocie}
\alias{bezrobocie}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
bezrobocie(x, sufiks, sufiksTlo = sufiks)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazwy zmiennej, na podstawie której ma
zostać obliczony wskaźnik (użyta zostanie zmienna o nazwie:
"bezrobocie\code{sufiks}")}

\item{sufiksTlo}{opcjonalnie ciąg znaków - sufiks nazwy zmiennej
opisującej stopę bezrobocia w powiecie (użyta zostanie zmienna o nazwie:
"powiat_bezrobocie\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.nie
        mających braku danych w zmiennej "bezrobocie\code{sufiks}"),}
  \item{\code{`zarejestrowani bezrobotni`} - liczba [0, n),}
  \item{\code{`pracujący lub kontynuujący naukę`} - liczba [0, n).}
}
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę zarejestrowanych
bezrobotnych w grupie.
}
\seealso{
\code{\link{bezrobocie_miesiac}}
}
