% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{brak_pracy}
\alias{brak_pracy}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
brak_pracy(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca liczebności absolwentów, którzy nie
pracowali w momencie badania, w tym zamierzających podjąć pracę w ciągu
następnych 12 miesięcy, w tym planujących podjęcie pracy w wyuczonym
zawodzie.
}
