% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{czas_rozklad}
\alias{czas_rozklad}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
czas_rozklad(x, zmienna, punktyPodzialu, etykiety)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{zmienna}{ciąg znaków - nazwa zmiennej, na podstawie której ma być
obliczony wskaźnik}

\item{punktyPodzialu}{wektor liczb - punkty wyznaczające granice przedziałów,
w ramach których zliczane będą wystąpienia badanych}

\item{etykiety}{wektor ciągów znaków - etykiety przedziałów wyznaczonych
przez \code{granicePrzedzialow}}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.nie
        mających braku danych w zmiennej opisanej argumentem \code{zmienna}),}
  \item{\code{`etykieta[1]`} - liczba [0, n),}
  \item{...}
}
}
\description{
Funkcja oblicza wskaźnik opisujący rozkład wskaźnika z poziomu
indywidualnego w ramach grupy - zwykle opisującego czas pracy lub bezrobocia
w badanym okresie. Wartości wskaźnika z poziomu indywidualnego przekształcane
są na zmienną porządkową poprzez pocięcie na przedziały zdefiniowane przez
argument \code{punktyPodzialu}, a następnie dla każdej z wartości takiej
zmiennej porządkowej zwracana jest liczba jej wystąpień w ramach grupy.
}
\seealso{
\code{\link{praca_czas}}, \code{\link{bezrobocie_czas}}
}
