% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{czy_plany_eduk}
\alias{czy_plany_eduk}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
czy_plany_eduk(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca procent absolwentów, którzy w ciągu 6
miesięcy od ukończenia szkoły planują rozpocząć dalszą naukę.
}
