% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{dane_szkoly}
\alias{dane_szkoly}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
dane_szkoly(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca nazwę i adres szkoły jako listę.
}
