% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{egz_zaw_zdawalnosc}
\alias{egz_zaw_zdawalnosc}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
egz_zaw_zdawalnosc(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}
}
\value{
lista o elementach:
\itemize{
  \item{\code{`n`} - liczba badanych uwzględnionych w obliczeniach (tj. nie
        mających braku danych w zmiennej \code{egz_zaw_zdany}),}
  \item{\code{`zdawalność`} - liczba [0, 1].}
}
}
\description{
Funkcja oblicza wskaźnik opisujący średnią zdawalność egzaminu
zawodowego w grupie.
}
\seealso{
\code{\link{wskazniki_nie_z_epizodow}}
}
