% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{firma_badawcza}
\alias{firma_badawcza}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
firma_badawcza(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
ciąg znaków
}
\description{
Funkcja przechowująca informację o nazwie firmy realizującej
badanie.
}
