% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{formy}
\alias{formy}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
formy(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca listę form gramatycznych różnych słów lub
wyrażeń, które pojawiają się w raporcie w zależności od typu szkoły.
}
