% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputuj_miesiac_pk_1rm.R
\name{imputuj_mies_pk_1rm}
\alias{imputuj_mies_pk_1rm}
\title{Imputacja miesiacow rozpoczecia i zakonczenia epizodow w danych sondazowych}
\usage{
imputuj_mies_pk_1rm(x)
}
\arguments{
\item{x}{lista zwracana przez funkcję \code{\link{wczytaj_wyniki_pilrm}}}
}
\value{
ramka danych
}
\description{
Funkcja imputuje braki danych w zmiennych opisujących miesiąc rozpoczęcia
i miesiąc zakończenia epizodów pracy nauki i bezrobocia w danych sondażowych
z pilotażowej rundy monitringu.
}
