% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputuj_miesiac_pk_pilrm.R
\name{imputuj_miesiac_pk_pilrm}
\alias{imputuj_miesiac_pk_pilrm}
\title{Imputacja miesiacow rozpoczecia i zakonczenia epizodow w danych sondazowych}
\usage{
imputuj_miesiac_pk_pilrm(x)
}
\arguments{
\item{x}{lista zwracana przez funkcję \code{\link{wczytaj_wyniki_pilrm}}}
}
\value{
lista ramek danych o takiej samej strukturze, jak zwracana przez
\code{\link{wczytaj_wyniki_pilrm}}
}
\description{
Funkcja imputuje braki danych w zmiennych opisujących miesiąc rozpoczęcia
i miesiąc zakończenia epizodów pracy nauki i bezrobocia w danych sondażowych
z pilotażowej rundy monitringu.
}
\details{
Patrz dokumentacja dot. struktury i zawartości zbiorów danych
z \href{../doc/runda_pilotazowa-dokumentacja.html}{rundy pilotażowej}
tworzonych przy pomocy funkcji pakietu MLASZdane.
}
