% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koryguj_statusy.R
\name{koryguj_statusy}
\alias{koryguj_statusy}
\title{Tworzenie zbioru osobo-miesiecy}
\usage{
koryguj_statusy(
  x,
  zmiennaID,
  kodyPracyNaCzarno = vector(mode = "numeric", length = 0)
)
}
\arguments{
\item{x}{ramka danych obemująca opis statusów \strong{jednego} badanego}

\item{zmiennaID}{ciąg znaków - nazwa zmiennej przechowującej unikalne ID badanego}

\item{kodyPracyNaCzarno}{wektor liczbowy - wartości zmiennej \code{praca},
które opisują pracę na czarno (czyli takie, które nie wchodzą w formalny
konflikt z posiadaniem statusu zarejestrwoanego bezrobotnego)}
}
\value{
ramka danych ze skorygowanymi statusami i dopisanymi kolumnami
opisującymi, jakie zmiany zostały dokonane
}
\description{
Funkcja odpowiada za rozstrzyganie konfliktów pomiędzy statusami
nauki/pracy/bezrobocia w poszczególnych miesiącach wynikającymi z deklaracji
respondentów. Jest wykorzystywana wewnątrz wywołań funkcji
\code{\link{przygotuj_zbior_osobo_miesiecy_pilrm}}
i \code{\link{przygotuj_zbior_osobo_miesiecy_1rm}}.
}
