% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kzsb}
\alias{kzsb}
\title{Klasyfikacja Zawodow Szkolnictwa Branzowego.}
\format{
Ramka danych z 19 kolumnami i 261 wierszami.
}
\usage{
kzsb
}
\description{
Klasyfikacja Zawodów Szkolnictwa Branżwowego zawiera
w szczególności kody zawodów z przypisanymi im nazwami branż wprowadzonych
rozporządzeniem MEN w 2019 roku. Do jednej branży przypisany są najczęściej
więcej niż jeden kod zawodu.
}
\keyword{datasets}
