% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{liczba_kobiet_2rm}
\alias{liczba_kobiet_2rm}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
liczba_kobiet_2rm(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
liczba
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę zbadanych kobiet w
grupie.
}
