% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R,
%   R/wskazniki_poziom_zagregowany_2rm.R
\name{liczba_zbadanych}
\alias{liczba_zbadanych}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
liczba_zbadanych(x)

liczba_zbadanych(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
liczba

liczba
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę zbadanych w grupie.

Funkcja oblicza wskaźnik opisujący liczbę zbadanych w grupie.
}
