% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{mlodociani_praca}
\alias{mlodociani_praca}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
mlodociani_praca(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca liczebności absolwentów będących
młodocianymi pracownikami, w tym zamierzjącymi podjąć pracę oraz
przechowująca liczbę młodocianych pracowników planujących pracę w wyuczonym
zawodzie.
}
