% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{nauka}
\alias{nauka}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
nauka(x, sufiks)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazwy zmiennej, na podstawie której ma
zostać obliczony wskaźnik (użyta zostanie zmienna o nazwie:
"nauka\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.nie
        mających braku danych w zmiennej "nauka\code{sufiks}"),}
  \item{\code{`studia stacjonarne`} - liczba [0, n],}
  \item{\code{`studia niestacjonarne`} - liczba [0, n],}
  \item{\code{`szkoła policealna`} - liczba [0, n],}
  \item{\code{`LO dla dorosłych`} - liczba [0, n].}
}
}
\description{
Funkcja oblicza wskaźnik opisujący rozkład formy dalszej
edukacji w grupie.
}
\seealso{
\code{\link{nauka_miesiac}}
}
