% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{nauka_czas}
\alias{nauka_czas}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
nauka_czas(epizody, limitG, limitD, idAbsolwenta = "ID_RESP")
}
\arguments{
\item{epizody}{ramka danych z epizodami - typowo element \code{epizody} listy
zwracanej przez funkcję \code{\link{imputuj_miesiac_pk_1rm}}}

\item{limitG}{liczba - górna granica, do jakiej zostanie zawężony analizowany
okres, wyrażona jako liczba miesięcy od momentu planowego ukończenia szkoły
(porównywana z wartością kolumny \code{czas_rozp})}

\item{limitD}{liczba - dolna granica, do jakiej zostanie zawężony analizowany
okres, wyrażona jako liczba miesięcy od momentu planowego ukończenia szkoły
(porównywana z wartością kolumny \code{czas_zakon})}

\item{idAbsolwenta}{nazwa zmiennej identyfikującej badanych (podana jako ciąg
znaków lub wyrażenie)}
}
\value{
data frame
}
\description{
Na podstawie zbioru epizodów dla każdego badanego funkcja
zwraca, w których miesiącach się on uczył. Opisująca to w zwracanej ramce
danych zmienna \code{nauka_czas} jest kolumną-listą.
}
\details{
Funkcja wykorzystywana jest wewnętrznie w ramach przekształceń
dokonywanych przez funkcje \code{\link{praca_czas}}
i \code{\link{bezrobocie_czas}}.
}
