% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{nauka_zawod}
\alias{nauka_zawod}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
nauka_zawod(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca rozkład odpowiedzi na pytanie PNZ9 "Jak
Pani/Pan ocenia, czy nauka u tego pracodawcy przygotowała P. do wykonywania
pracy w zawodzie, którego się Pan(i) uczył(a)?" oraz sumę odsetków
odpowiedzi: "Przygotowała doskonale" i "Przygotowała dobrze".
}
