% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nazwy_jst}
\alias{nazwy_jst}
\title{Nazwy powiatow i wojewodztw przypisane do kodow TERYT powiatow.}
\format{
Ramka danych z 3 kolumnami i 380 wierszami.
}
\usage{
nazwy_jst
}
\description{
Ramka danych z czterocyforwymi (lub trzycyfrowymi) kodami TERYT
oraz z odpowiadającymi im nazwami powiatów oraz województw.
}
\keyword{datasets}
