% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{oblicz_wskazniki_ind_1rm}
\alias{oblicz_wskazniki_ind_1rm}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
oblicz_wskazniki_ind_1rm(x, idAbsolwenta = "ID_RESP")
}
\arguments{
\item{x}{ramka danych zwracana przez funkcję
\code{\link{imputuj_miesiac_pk_1rm}}}

\item{idAbsolwenta}{nazwa zmiennej identyfikującej badanych (podana jako ciąg
znaków lub wyrażenie)}
}
\value{
data frame zawierający następujące kolumny:
\itemize{
  \item{wszystkie kolumny ramki danych pr}
}
}
\description{
Funkcja oblicza zmienne-wskaźniki, które wymagają informacji
nie ze zbioru epizodów.
}
