% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{oblicz_wskazniki_w_forku}
\alias{oblicz_wskazniki_w_forku}
\title{Obliczanie wskaznikow na poziomie zagregowanym - funkcje pomocnicze}
\usage{
oblicz_wskazniki_w_forku(grupa, .data, funkcje)
}
\arguments{
\item{grupa}{wyrażenie zawierające definicję grupy (lub grupy odniesienia)}

\item{.data}{ramka danych (zawierająca wskaźniki na poziomie indywidualnym)}

\item{funkcje}{lista wyrażeń przekazanych w wywołaniu
\code{\link{agreguj_wskazniki}} poprzez \code{...}}
}
\value{
ramka danych
}
\description{
Nieeksportowana funkcja używana w ramach
\code{\link{agreguj_wskazniki}} do obliczania wsartości wskaźników, kiedy
wykorzystywana jest wielowątkowość w systemach innych niż Windows (tj.,
kiedy możliwe jest tworzenie \emph{forków} procesów)
}
\seealso{
\code{\link{agreguj_wskazniki}}
}
