% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{oblicz_wskazniki_w_watku}
\alias{oblicz_wskazniki_w_watku}
\title{Obliczanie wskaznikow na poziomie zagregowanym - funkcje pomocnicze}
\usage{
oblicz_wskazniki_w_watku(.data, funkcje)
}
\arguments{
\item{.data}{ramka danych (zawierająca wskaźniki na poziomie indywidualnym)}

\item{funkcje}{lista wyrażeń przekazanych w wywołaniu
\code{\link{agreguj_wskazniki}} poprzez \code{...}}
}
\value{
ramka danych
}
\description{
Nieeksportowana funkcja używana w ramach
\code{\link{agreguj_wskazniki}} do obliczania wsartości wskaźników, kiedy
wykorzystywana jest wielowątkowość w systemach Windows (tj. z użyciem
\emph{socket connections})
}
\seealso{
\code{\link{agreguj_wskazniki}}
}
