% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{ocena_pnz}
\alias{ocena_pnz}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
ocena_pnz(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca informację o zadowoleniu z praktycznej
nauki zawodu (PNZ). Definicja wskaźnika: % PNZ3, PNZ5, PNZ7, PNZ9 - co
najmniej jedna odpowiedź “przygotowała dobrze” lub “przygotowała bardzo
dobrze”
}
