% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{placa_brutto}
\alias{placa_brutto}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
placa_brutto(x, przelicznik = 0.71)
}
\arguments{
\item{x}{wysokość zarobków wyrażona w kwocie netto - zmienna
\strong{pio4pelen} z obiektu \code{epizody}}

\item{przelicznik}{przyjęty odgórnie przelicznik kwoty netto na kwotę brutto.
Posiada domyślną wartość 0.71}
}
\description{
Prosta funkcja przeliczająca wysokość zarobków wyrażoną w kwocie
netto na kwotę brutto. Przyjęty został przybliżony przeliczniok o wartości
0.71
}
