% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{plany_6m}
\alias{plany_6m}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
plany_6m(x, zmienna, opis)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}

\item{zmienna}{zmienna znajdująca się w ramce danych \code{x}. Funkcja służy
do wyliczania wskaźników na zmiennych \code{PL2_1}, \code{PL2_2} oraz
\code{PL2_4}.}

\item{opis}{opis zwykle odnoszący się do argumentu przekazanego w
\code{zmienna}}
}
\value{
lista
}
\description{
Funkcja przechowująca sumę wskazań odpowiedzi "Raczej tak" oraz
"Zdecydowanie tak" w pytaniu PL2 "Czy w ciągu 6 miesięcy po zakończeniu nauki
w obecnej szkole planuje Pani/Pan rozpocząć naukę" dla 1., 2. i 4. pozycji w
kafeterii.
}
