% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{plany_eduk_nie}
\alias{plany_eduk_nie}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
plany_eduk_nie(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca procent absolwentów, którzy albo nie pracują
w wyuczonym zawodzie albo nie planują w ciągu najbliższych 12 miesięcy
rozpoczynać pracy w wyuczonym zawodzie.
}
