% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{plany_eduk_tak}
\alias{plany_eduk_tak}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
plany_eduk_tak(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca procent absolwentów, którzy albo pracują w
wyuczonym zawodzie albo planują w ciągu najbliższych 12 miesięcy rozpocząć
pracę w wyuczonym zawodzie.
}
