% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl.R
\name{pobierz_dane_bdl}
\alias{pobierz_dane_bdl}
\title{Pobieranie danych z Banku Danych Lokalnych GUS}
\usage{
pobierz_dane_bdl(wskazniki, lata, poziom = "powiaty")
}
\arguments{
\item{wskazniki}{data frame zwrócony przez funkcję
\code{\link{znajdz_wskazniki_bdl}} lub \code{\link{wskaznik_bdl}}}

\item{lata}{wektor liczbowy - lata, dla których mają zostać zwrócone
wartości wskaźników}

\item{poziom}{opcjonalnie poziom agregacji wskaźników: "makroregiony",
"województwa", "regiony", "podregiony", "powiaty" lub "gminy"}
}
\value{
ramka danych z przypisaną klasą "daneBDL"
}
\description{
Funkcja przy pomocy API BDL pobiera zestawienie wartości podanych wskaźników
w podanych latach na podanym poziomie agregacji.
}
\details{
\strong{Uwaga!} API BDL ma absurdalnie niskie limity na dopuszczalną
w przedziale czasu liczbę zapytań (a przy tym ogranicza liczbę zwracanych
wierszy zestawienia do maksimum 100), więc pobieranie wskaźników przy pomocy
tej funkcji należy sobie dobrze rozplanować w czasie. W związku z tym zasadne
jest też zapisanie ściągniętych zestawień lokalnie do wykorzystania
w przyszłości, bez konieczności używania API.
}
\seealso{
\code{\link{przeksztalc_dane_bdl}}
}
