% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokaz_ramke_danych.R
\name{pokaz_ramke_danych}
\alias{pokaz_ramke_danych}
\title{Funkcje pomocnicze}
\usage{
pokaz_ramke_danych(x, sep = " ")
}
\arguments{
\item{x}{ramka danych}

\item{sep}{ciąg znaków - separator poziomy pól}
}
\value{
ciąg znaków
}
\description{
Nieeksportowana funkcja zwracająca zawartość ramki danych
w formie ciągu znaków, który można wydrykować przy pomocy np. funkcji
\code{\link[base]{message}}, zamiast używać \code{print}
}
\seealso{
\code{\link{agreguj_wskazniki}}
}
