% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polacz_dane.R
\name{polacz_dane}
\alias{polacz_dane}
\title{laczy dane ZUS, o wynikach egzaminow, uczniach, studentach i szkolach
oraz statystyki powiatow}
\usage{
polacz_dane(absolwenci, zus, zdajacyEgzaminy, uczniowie, studenci, szkoly,
  pnaPowiaty, dataMin, dataMax)
}
\arguments{
\item{absolwenci}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_absolwenci}}}

\item{zus}{dane wygenerowane za pomocą funkcji
\code{\link[MLAKdane]{przygotuj_zus}}}

\item{zdajacyEgzaminy}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_zdajacy}}}

\item{uczniowie}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_uczniowie}}}

\item{studenci}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_studenci}}}

\item{szkoly}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_szkoly}}}

\item{pnaPowiaty}{dane wygenerowane za pomocą funkcji
\code{\link[MLAKdane]{przygotuj_pna_powiaty_mb}}}

\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch
znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch
znaków, np. '2015-09-30')}
}
\value{
data.frame - połączone dane
}
\description{
Aby poprawnie przyłączyć statystyki powiatów macierz danych
uzupełniana jest tak, aby dla każdego absolwenta zawierała informacje o
każdym okresie w zadanym przedziale czasu, ale nie później niż do "końca
obserwacji w ZUS" (zakładając zamieszkanie absolwenta w uzupełnianych
okresach "w Polsce", kodowane jako pna = -1, oraz wszelkie cechy tytułu
ubezpieczenia ZUS równe 0, w szczególności także bezrob = 0).
}
