% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{ponowny_wybor}
\alias{ponowny_wybor}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
ponowny_wybor(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca rozkład odpowiedzi na pytanie W2 "Gdyby
teraz mógł (mogła) Pani/Pan wybrać szkołę, to czy ponownie wybrał(a)by P. tę
samą szkołę, w której obecnie się P. uczy?" oraz odsetek sum odpowiedzi
"Raczej tak" i "Na pewno tak".
}
