% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{praca_czas_rozp}
\alias{praca_czas_rozp}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_czas_rozp(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}
}
\value{
lista o elementach:
\itemize{
  \item{\code{`n`} - liczba badanych uwzględnionych w obliczeniach (tj. nie
        mających braku danych w zmiennej \code{praca_czas_rozp_pierwsza}),}
  \item{\code{`średnia`} - liczba [0, ),}
  \item{\code{`mediana`} - liczba [0, ),}
  \item{\code{`1.kwartyl`} - liczba [0, )}
  \item{\code{`3.kwartyl`} - liczba [0, ).}
}
}
\description{
Funkcja oblicza wskaźnik opisujący parametry rozkładu czasu
od momentu planowego ukończenia szkoły do podjęcia pierwszej pracy
(wśród badanych, którzy nie pracowali przed ukończeniem szkoły).
}
\seealso{
\code{\link{praca_pierwsza}}
}
