% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{praca_forma}
\alias{praca_forma}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_forma(x, sufiks)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazw zmiennych, na podstawie których ma
zostać obliczony wskaźnik (użyte zostaną zmienne o nazwach:
"pg2h([123456789]|10).\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.
        pracujących, z wyłączeniem pracujących bez umowy),}
  \item{\code{`umowa o pracę na czas nieokreślony`} - liczba [0, n],}
  \item{\code{`umowa o pracę na czas określony`} - liczba [0, n],}
  \item{\code{`umowa zlecenia/o dzieło`} - liczba [0, n],}
  \item{\code{`samozatrudnienie`} - liczba [0, n],}
  \item{\code{`rolnicy indywidualni`} - liczba [0, n],}
  \item{\code{`przez agencję pracy tymczasowej`} - liczba [0, n],}
  \item{\code{`staż lub praktyka`} - liczba [0, n],}
  \item{\code{`inna (lub nieznana)`} - liczba [0, n].}
}
}
\description{
Funkcja oblicza wskaźnik opisujący rozkład formy zatrudnienia
w grupie.
}
\seealso{
\code{\link{praca_miesiac}}, \code{\link{praca_pierwsza}},
\code{\link{praca_ostatnia}}, \code{\link{praca_forma2}}
}
