% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{praca_miesiac}
\alias{praca_miesiac}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
praca_miesiac(epizody, miesiac, idAbsolwenta = "ID_RESP")
}
\arguments{
\item{epizody}{ramka danych z epizodami - typowo element \code{epizody} listy
zwracanej przez funkcję \code{\link{imputuj_miesiac_pk_1rm}}}

\item{miesiac}{liczba - miesiąc od terminu planowego ukończenia szkoły
(porównywana z wartościami kolumn \code{czas_rozp} i \code{czas_zakon})}

\item{idAbsolwenta}{nazwa zmiennej identyfikującej badanych (podana jako ciąg
znaków lub wyrażenie)}
}
\value{
data frame
}
\description{
Funkcja oblicza zmienne-wskaźniki opisujące formę (formy)
zatrudnienia i miejsce (miejsca) pracy (w stosunku do miejsca zamieszkania)
badanego w określonym miesiącu od terminu planowego ukończenia szkoły. Nie
jest przy tym brana pod uwagę praca świadczona bez umowy.
}
\seealso{
\code{\link{praca_forma}}, \code{\link{praca_forma2}},
\code{\link{praca_zamieszkanie}}
}
