% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{praca_nauka}
\alias{praca_nauka}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_nauka(x, sufiks)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazwy zmiennej, na podstawie której ma
zostać obliczony wskaźnik (użyta zostanie zmienna o nazwie:
"praca_nauka\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{`n`} - liczba wszystkich badanych w grupie,}
  \item{\code{`tylko pracują`} - liczba [0, n],}
  \item{\code{`pracują i uczą się`} - liczba [0, n],}
  \item{\code{`tylko się uczą`} - liczba [0, n],}
  \item{\code{`nie pracują i nie uczą się`} - liczba [0, n].}
}
}
\description{
Funkcja oblicza wskaźnik opisujący rozkład statusu
edukacyjno-zawodowego w grupie.
}
\seealso{
\code{\link{praca_nauka_miesiac}}
}
