% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{praca_nauka_miesiac}
\alias{praca_nauka_miesiac}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
praca_nauka_miesiac(epizody, miesiac, idAbsolwenta = "ID_RESP")
}
\arguments{
\item{epizody}{ramka danych z epizodami - typowo element \code{epizody} listy
zwracanej przez funkcję \code{\link{imputuj_miesiac_pk_1rm}}}

\item{miesiac}{liczba - miesiąc od terminu planowego ukończenia szkoły
(porównywana z wartościami kolumn \code{czas_rozp} i \code{czas_zakon})}

\item{idAbsolwenta}{nazwa zmiennej identyfikującej badanych (podana jako ciąg
znaków lub wyrażenie)}
}
\value{
data frame
}
\description{
Funkcja oblicza zmienną-wskaźnik opisującą status
edukacyjno-zawodowy respondenta w podanym miesiącu od momentu planowego
ukończenia szkoły.
}
\seealso{
\code{\link{praca_nauka}}
}
