% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{praca_pierwsza}
\alias{praca_pierwsza}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
praca_pierwsza(epizody, idAbsolwenta = "ID_RESP")
}
\arguments{
\item{epizody}{ramka danych z epizodami - typowo element \code{epizody} listy
zwracanej przez funkcję \code{\link{imputuj_miesiac_pk_1rm}}}

\item{idAbsolwenta}{nazwa zmiennej identyfikującej badanych (podana jako ciąg
znaków lub wyrażenie)}
}
\value{
data frame
}
\description{
Funkcja oblicza zmienne-wskaźniki opisujące pierwszą pracę
badanego wykonywaną po ukończeniu szkoły (również jeśli zaczęła się
wcześniej). Nie jest przy tym brana pod uwagę praca świadczona bez umowy.
}
\seealso{
\code{\link{praca_forma_ind}}, \code{\link{praca_zamieszkanie_ind}},
\code{\link{praca_forma}}, \code{\link{praca_forma2}},
\code{\link{praca_przed_ukonczeniem_szkoly}},
\code{\link{praca_zamieszkanie}},
\code{\link{praca_zgodna_z_wyksztalceniem}},
\code{\link{praca_zarobki}}
}
