% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{praca_poza_wyuczonym}
\alias{praca_poza_wyuczonym}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_poza_wyuczonym(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca liczebności absolwentów, którzy pracują
poza wyuczonym zawodem, w tym zamierzących podjąć pracę w wyuczonym zawodzie
w ciągu następnych 12 miesięcy.
}
