% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{praca_przed_ukonczeniem_szkoly}
\alias{praca_przed_ukonczeniem_szkoly}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_przed_ukonczeniem_szkoly(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}
}
\value{
lista o elementach:
\itemize{
  \item{\code{`n`} - liczba badanych w grupie,}
  \item{\code{`praca przed ukończeniem szkoły`} - liczba [0, n].}
}
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę absolwentów w grupie,
którzy pracowali przed ukończeniem (badanej) szkoły.
}
\seealso{
\code{\link{praca_pierwsza}}
}
