% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{praca_spelnienie_oczekiwan}
\alias{praca_spelnienie_oczekiwan}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_spelnienie_oczekiwan(x, sufiks)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazwy zmiennej, na podstawie której ma
zostać obliczony wskaźnik (użyta zostanie zmienna o nazwie:
"po5\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.nie
        mających braku danych w zmiennej "po5\code{sufiks}"),}
  \item{\code{`niedostateczny`} - liczba [0, n],}
  \item{\code{`dopuszczający`} - liczba [0, n],}
  \item{\code{`dostateczny`} - liczba [0, n],}
  \item{\code{`dobry`} - liczba [0, n],}
  \item{\code{`bardzo dobry`} - liczba [0, n],}
  \item{\code{`celujący`} - liczba [0, n].}
}
}
\description{
Funkcja oblicza wskaźnik opisujący rozkład zadowolenia
z ostatniej pracy w analizowanym okresie.
}
\seealso{
\code{\link{praca_pierwsza}}, \code{\link{praca_ostatnia}}
}
