% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{praca_w_zawodzie}
\alias{praca_w_zawodzie}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_w_zawodzie(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca liczebności absolwentów ze względu na
status zawowy (praca zarobkowa, brak pracy zarobkowej, praca poza wyuczonym
zawodem, młodociani pracownicy).
}
