% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{praca_zamieszkanie}
\alias{praca_zamieszkanie}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_zamieszkanie(x, sufiks)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazw zmiennych, na podstawie których ma
zostać obliczony wskaźnik (użyte zostaną zmienne o nazwach:
"pg2i[123].\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.
        pracujących, z wyłączeniem pracujących bez umowy),}
  \item{\code{`w Polsce`} - liczba [0, n],}
  \item{\code{`za granicą`} - liczba [0, n],}
  \item{\code{`nieznane`} - liczba [0, n].}
}
}
\description{
Funkcja oblicza wskaźnik opisujący rozkład miejsca zamieszkania
(w kraju czy za granicą) badanych, którzy pracowali.
}
\seealso{
\code{\link{praca_miesiac}}, \code{\link{praca_pierwsza}},
\code{\link{praca_ostatnia}}
}
