% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{praca_zamieszkanie_ind}
\alias{praca_zamieszkanie_ind}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
praca_zamieszkanie_ind(x, idAbsolwenta = "ID_RESP")
}
\arguments{
\item{x}{ramka danych z epizodami - typowo element \code{epizody} listy
zwracanej przez funkcję \code{\link{imputuj_miesiac_pk_1rm}}}

\item{idAbsolwenta}{nazwa zmiennej identyfikującej badanych (podana jako ciąg
znaków lub wyrażenie)}
}
\value{
data frame
}
\description{
Funkcja oblicza zmienne-wskaźniki opisujące, czy badany
kiedykolwiek był zatrudniony w miejscu innym, niż miejscowość zamieszkania
}
\seealso{
\code{\link{praca_miesiac}}, \code{\link{praca_pierwsza}},
\code{\link{praca_ostatnia}}
}
