% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{praca_zarobki}
\alias{praca_zarobki}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_zarobki(x, sufiks, sufiksTlo = NULL)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazwy zmiennej, na podstawie której ma
zostać obliczony wskaźnik (użyta zostanie zmienna o nazwie:
"pio4\code{sufiks}")}

\item{sufiksTlo}{opcjonalnie ciąg znaków - sufiks nazwy zmiennej
opisującej średnie zarobki w powiecie (użyta zostanie zmienna o nazwie:
"powiat_sr_wynagrodzenia\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.nie
        mających braku danych w zmiennej "po5\code{sufiks}"),}
  \item{\code{`średnia`} - liczba [0, ),}
  \item{\code{`mediana`} - liczba [0, ),}
  \item{\code{`1.kwartyl`} - liczba [0, ),}
  \item{\code{`3.kwartyl`} - liczba [0, ),}
  \item{\code{`tlo`} - liczba [0, ).}
}
}
\description{
Funkcja oblicza wskaźnik opisujący rozkład zarobków w grupie.
}
\seealso{
\code{\link{praca_pierwsza}}, \code{\link{praca_ostatnia}}
}
