% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{praca_zarobkowa}
\alias{praca_zarobkowa}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_zarobkowa(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca liczebności absolwentów, którzy pracowali
zawodowo w momencie badania, w tym pracujących w wyuczonym zawodzie.
}
