% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{praca_zgodna_z_wyksztalceniem}
\alias{praca_zgodna_z_wyksztalceniem}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
praca_zgodna_z_wyksztalceniem(x, sufiks)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazw zmiennych, na podstawie których ma
zostać obliczony wskaźnik (użyte zostaną zmienne o nazwach:
"pg2i[123].\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.
        pracujących, z wyłączeniem pracujących bez umowy),}
  \item{\code{`zgodna`} - liczba [0, n],}
  \item{\code{`niezgodna, ale wymaga podobnych kwalifikacji`} - liczba [0, n],}
  \item{\code{`niezgodna, wymaga innych kwalifikacji`} - liczba [0, n],}
  \item{\code{`praca, w której wykształcenie nie ma znaczenia`} - liczba [0, n].}
}
}
\description{
Funkcja oblicza wskaźnik opisujący rozkład subiektywnej oceny
zgodności pierwszej lub ostatniej wykonywanej pracy z kierunkiem kształcenia.
}
\seealso{
\code{\link{praca_pierwsza}}, \code{\link{praca_ostatnia}}
}
