% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{proste_wskazniki_na_wektor}
\alias{proste_wskazniki_na_wektor}
\title{Obliczanie wskaznikow na poziomie zagregowanym - funkcje pomocnicze}
\usage{
proste_wskazniki_na_wektor(x)
}
\arguments{
\item{x}{ramka danych (zawierająca wskaźniki na poziomie zagregowanym)}
}
\value{
ramka danych
}
\description{
Nieeksportowana funkcja używana w ramach
\code{\link{agreguj_wskazniki}} do zamiany kolumn zawierających wskaźniki
które są (dla każdej grupy) pojedynczą liczbą z kolumn-list na wektory
}
\seealso{
\code{\link{agreguj_wskazniki}}
}
