% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl.R
\name{przeksztalc_dane_bdl}
\alias{przeksztalc_dane_bdl}
\title{Pobieranie danych z Banku Danych Lokalnych GUS}
\usage{
przeksztalc_dane_bdl(daneBdl, rocznik, prefiks = "")
}
\arguments{
\item{daneBdl}{data frame zwrócony przez funkcję \code{\link{pobierz_dane_bdl}}}

\item{rocznik}{liczba - rok ukończenia szkoły, w stosunku do którego mają
zostać skonstruowane nazwy wskaźników}

\item{prefiks}{opcjonalnie prefiks dopisywany na początku nazw (nazw,
nie wartości!) zmiennych opisujących JST (teryt i nazwę) w zwracanej ramce
danych}
}
\value{
ramka danych
}
\description{
Funkcja przetwarza dane pobrane z API BDL w formie \emph{długiej} (jedna
JST-wiele wierszy) na formę \emph{szeroką} (jedna JST-jeden wiersz), w której
są zdatne do wykorzystania przy obliczaniu wskaźników szkół dla konkretnego
rocznika absolwentów.
}
\details{
\strong{Uwaga!} W tej chwili zmieniając nazwy wskaźników na nazwy zmiennych
funkcja stosuje następujące reguły:
\itemize{
  \item{nazwy wszystkich wskaźników zawierające ciąg znaków "bezrobocia"
        traktowane są jako jeden wskaźnik, w nazwach wynikowych zmiennych
        opisywany jako "bezrobocie";}
  \item{nazwy wszystkich wskaźników zawierające ciąg znaków "wynagrodzenia"
        traktowane są jako jeden wskaźnik, w nazwach wynikowych zmiennych
        opisywany jako "sr_wynagrodzenia";}
  \item{nazwy wszystkich pozostałych wskaźników pozostają bez zmian, z tym
        że znaki spacji zamieniane są w nich na znak "_".}
}
\strong{Nie ma gwarancji, że ww. reguły nie ulegną zmianie w przyszłości!}

Nazwy zmiennych zawierających wartości wskaźników w zwracanym zestawieniu są
postaci: \code{jst_wskaznik_[[:digit:]]+[mr]}, gdzie:
\itemize{
  \item{\code{wskaznik} to nazwa wskaźnika (p. wyżej);}
  \item{\code{[[:digit:]]+} to liczba (miesięcy lub lat od momentu planowego
        ukończenia szkoły przez uczniów rocznika podanego w wywołaniu funkcji
        argumentem  \code{rocznik});}
  \item{\code{[mr]+} to litera "m" lub litera "r" w zależności od tego, czy
        dany wskaźnik raportowany jest przez GUS z częstotliwością miesięczną,
        czy roczną.}
}
Jeśli wszystkie wskaźniki w zestawieniu pobranym z API BDL przekazanym do
funkcji są określone na tym samym poziomie agregacji, to w nazwach zmiennych
ciąg znaków "jst" jest zamieniany na nazwę danego poziomu (np. "powiat").
}
