% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{przyg_zaw_prakt_ANY}
\alias{przyg_zaw_prakt_ANY}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
przyg_zaw_prakt_ANY(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca odsetek absolwentów, których praktyki
przygotowały "Przygotowały bardzo dobrze" i "Przygotowały dobrze" (pytanie
PNZ3 "Jak Pan(i) ocenia, czy praktyki przygotowały Panią/Pana do wykonywania
pracy w zawodzie, którego się P. uczył(a)?").
}
