% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{przyg_zawodu_prakt_PL}
\alias{przyg_zawodu_prakt_PL}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
przyg_zawodu_prakt_PL(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca rozkład odpowiedzi na pytanie PNZ3 "Jak
Pan(i) ocenia, czy praktyki przygotowały Panią/Pana do wykonywania pracy w
zawodzie, którego się P. uczył(a)? - U pracodawcy w Polsce" oraz sumę
odsetków wskazań odpowiedzi "Przygotowały bardzo dobrze" i "Przygotowały
dobrze".
}
