% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{przyg_zawodu_zaj_ckp}
\alias{przyg_zawodu_zaj_ckp}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
przyg_zawodu_zaj_ckp(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca rozkład odpowiedzi na pytanie PNZ5 "Jak P.
ocenia, czy zajęcia praktyczne przygotowały P. do wykonywania pracy w
zawodzie, którego się Pani/Pan uczył(a)? - W Centrum Kształcenia
Praktycznego/Zawodowego" oraz sumę odsetków wskazań odpowiedzi"Przygotowały
bardzo dobrze" i "Przygotowały dobrze".
}
