% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{przygotowanie_do_zawodu}
\alias{przygotowanie_do_zawodu}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
przygotowanie_do_zawodu(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca rozkład odpowiedzi na pytanie W3 "Jak P.
uważa, na ile szkoła przygotowała Panią/Pana do wykonywania zawodu, którego
się P. w niej uczy?" oraz sumy odsetków odpowiedzi: "Przygotowała doskonale"
i "Przygotowała dobrze" oraz "Nie przygotowała w ogóle" i "Przygotowała
słabo".
}
