% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_absolwenci.R
\name{przygotuj_absolwenci}
\alias{przygotuj_absolwenci}
\title{Przygotowuje plik absolwenci}
\usage{
przygotuj_absolwenci()
}
\value{
ramka danych (klasy 'absolwenci_df')
}
\description{
absolwenci to plik zawierający informacje o absolwentach szkół poddanych
badaniu monitorującemu - w szczególności zawierający informacje o tym, czy
wyrazili oni zgodę na wykorzystanie ich nr. PESEL do łączenia danych
(i ew. ten PESEL, jub jego zpseudonimizowane przekształcenie)
}
