% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_studenci.R
\name{przygotuj_studenci}
\alias{przygotuj_studenci}
\title{Przygotowuje plik studenci}
\usage{
przygotuj_studenci()
}
\value{
ramka danych (klasy 'studenci_df')
}
\description{
studenci to plik wyeksportowany z systemu POL-on, zawierający informacje
o absolwentach podlegających monitorowaniu, którzy kontunuują naukę na
uczelniach wyższych
}
