% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_szkoly.R
\name{przygotuj_szkoly}
\alias{przygotuj_szkoly}
\title{Przygotowuje plik szkoly}
\usage{
przygotuj_szkoly()
}
\value{
ramka danych (klasy 'szkoly_df')
}
\description{
szkoly to plik wyeksporotwany z SIO, zawierający informacje o szkołach,
dla których mają zostać obliczone wartości wskaźników
}
