% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_uczniowie.R
\name{przygotuj_uczniowie}
\alias{przygotuj_uczniowie}
\title{Przygotowuje plik uczniowie}
\usage{
przygotuj_uczniowie()
}
\value{
ramka danych (klasy 'uczniowie_df')
}
\description{
uczniowie to plik wyeksportowany z systemu SIO2, który zawiera informacje
o monitorowanych absolwentach, którzy kontynuują naukę w szkołach
(niebędących uczelniami wyższymi)
}
