% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splaszcz_wskazniki_zagregowane.R
\name{przygotuj_wskaznik_do_splaszczenia}
\alias{przygotuj_wskaznik_do_splaszczenia}
\title{Przeksztalca zbior wksaznikow z poziomu zagregowanego na plaska tabele}
\usage{
przygotuj_wskaznik_do_splaszczenia(x)
}
\arguments{
\item{x}{lista (zwykle: lista list o tej samej strukturze)}
}
\value{
ramka danych
}
\description{
Funkcja przekształca kolumnę-listę ramki danych w ramkę danych z wieloma
kolumnami będącymi zwykłymi wektorami (o liczbie wierszy równej długości
przekazanej listy).
}
