% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_zdajacy.R
\name{przygotuj_zdajacy}
\alias{przygotuj_zdajacy}
\title{Przygotowuje plik zdajacy_egzaminy}
\usage{
przygotuj_zdajacy()
}
\value{
ramka danych (klasy 'zdajacy_egzaminy_df')
}
\description{
zdajacy_egzaminy to plik z CKE/OKE, zawierający wyniki egzaminów zawodowych
monitorowanych absolwentów
}
