% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przywroc_etykiety.R
\name{przywroc_etykiety}
\alias{przywroc_etykiety}
\title{Obrobka danych etykietowanych}
\usage{
przywroc_etykiety(x, labTemp)
}
\arguments{
\item{x}{ramka danych, której zmiennym mają zostać przypisane etykiety}

\item{labTemp}{ramka danych (lub lista) z której mają zostać skopiowane
etykiety zmiennych i wartości zmiennych}
}
\value{
ramka danych
}
\description{
Funkcja pozwala szybko i wygodnie przywracać etykiety zmiennych i wartości,
które "gubią się" w niektórych operacjach na ramkach danych (o ile wcześniej
zostały one zapisane).
}
