% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splaszcz_wskazniki_zagregowane.R
\name{splaszcz_wskazniki_zagregowane}
\alias{splaszcz_wskazniki_zagregowane}
\title{Przeksztalca zbior wksaznikow z poziomu zagregowanego na plaska tabele}
\usage{
splaszcz_wskazniki_zagregowane(wskazniki)
}
\arguments{
\item{wskazniki}{ramka danych ze wskaźnikami na poziomie zagregowanym}
}
\value{
ramka danych
}
\description{
Funkcja przekształca zbiór danych ze wskaźnikami na poziomie zagregowanym,
w którym poszczególne wskaźniki mają postać kolumn-list na płaską ramkę
danych, w której każdy wskaźnik (który miał złożoną strukturę) reprezentowany
jest przez kilka różnych kolumn.
}
\details{
Większość pracy wykonuje pod spodem funkcja
\code{\link{przygotuj_wskaznik_do_splaszczenia}}. Następnie kolumny
konwertowane są na wektory typu "labelled", aby można im było przypisać
etykiety w bardzie opisowy sposób (niż poprawna leksykalnie nazwa zmiennej)
zdawały sprawę ze znaczenia zmiennych i aby etykiety takie bezproblemowo
zapisywały się przy eksporcie spłaszczonego zbioru danych do plików SPSS lub
Staty.
}
