% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{sprawdz_definicje_grup}
\alias{sprawdz_definicje_grup}
\title{Obliczanie wskaznikow na poziomie zagregowanym - funkcje pomocnicze}
\usage{
sprawdz_definicje_grup(wskazniki, definicje, dopuscPuste = FALSE)
}
\arguments{
\item{wskazniki}{ramka danych ze wskaźnikami na poziomie idywidualnym}

\item{definicje}{wektor lub lista zawierająca definicje podziałów na grupy}

\item{dopuscPuste}{wartość logiczna wskazująca, czy jako definicja grupy
dopuszczony jest brak danych - co może być używane w przypadku grup
odniesienia i oznacza brak grupy odniesienia}
}
\value{
wektor tekstowy o długości równej długości \code{definicje}:
zawiera komunikaty o błędach (jeśli wystąpiły) lub puste ciągi znaków (jeśli
definicja zewaluowała się bez problemu)
}
\description{
Nieeksportowana funkcja używana w ramach
\code{\link{agreguj_wskazniki}} do weryfikacji, czy wyrażenia definiujące
grupy, dla których mają zostać obliczone (zagregowane) wskaźniki, dają się
poprawnie zewaluować (w kontekście ramki danych ze wskaźnikami z poziomu
indywidualnego).
}
\seealso{
\code{\link{agreguj_wskazniki}}
}
