% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawdz_nazwy.R
\name{sprawdz_nazwy}
\alias{sprawdz_nazwy}
\title{Kontrola kompletnosci zbioru wskaznikow}
\usage{
sprawdz_nazwy(nazwy, wzorzec)
}
\arguments{
\item{nazwy}{wektor tekstowy - nazwy, których kompletność ma być sprawdzona}

\item{wzorzec}{wektor tekstowy - nazwy, z których każda powinna wystąpić
w \code{nazwy}}
}
\description{
Funkcja sprawdza, czy w podanym zestawie nazw występują
wszystkie wymagane nazwy.
}
