% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{studia_gdzie}
\alias{studia_gdzie}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
studia_gdzie(x, sufiks)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazwy zmiennej, na podstawie której ma
zostać obliczony wskaźnik (użyta zostanie zmienna o nazwie:
"studia_(kierunek|uczelnia)\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.nie
        mających braku danych w zmiennej "studia_uczelnia\code{sufiks}"),}
  \item{\code{`najczęściej wybierane kierunki`} - ciąg znaków - nazwy
        kierunków, oddzielone przecinkami, ułożone w kolejności od
        najczęściej do najrzadziej wystepujących, z pominięciem tych,
        których częstość występowania była mniejsza niż 20\%,}
  \item{\code{`najczęściej wybierane uczelnie`} - - ciąg znaków - nazwy
        uczelni, oddzielone przecinkami, ułożone w kolejności od
        najczęściej do najrzadziej wystepujących, z pominięciem tych,
        których częstość występowania była mniejsza niż 20\%.}
}
}
\description{
Funkcja oblicza wskaźnik opisujący kierunki i uczelnie, na
których kontuowali kształcenie badani z grupy.
}
\seealso{
\code{\link{studia_pierwsze}}
}
