% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{studia_pierwsze}
\alias{studia_pierwsze}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
studia_pierwsze(epizody, idAbsolwenta = "ID_RESP")
}
\arguments{
\item{epizody}{ramka danych z epizodami - typowo element \code{epizody} listy
zwracanej przez funkcję \code{\link{imputuj_miesiac_pk_1rm}}}

\item{idAbsolwenta}{nazwa zmiennej identyfikującej badanych (podana jako ciąg
znaków lub wyrażenie)}
}
\value{
data frame
}
\description{
Funkcja oblicza zmienne-wskaźniki opisujące pierwszą studia
podjęte przez badanego po ukończeniu szkoły.
}
\seealso{
\code{\link{studia_gdzie}}, \code{\link{studia_odplatnosc}},
\code{\link{studia_tryb}}
}
