% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{studia_tryb}
\alias{studia_tryb}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
studia_tryb(x, sufiks)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}

\item{sufiks}{ciąg znaków - sufiks nazwy zmiennej, na podstawie której ma
zostać obliczony wskaźnik (użyta zostanie zmienna o nazwie:
"studia_tryb\code{sufiks}")}
}
\value{
lista o elementach:
\itemize{
  \item{\code{n} - liczba badanych uwzględnionych w obliczeniach (tj.nie
        mających braku danych w zmiennej "studia_tryb\code{sufiks}"),}
  \item{\code{`studiujący na studiach dziennych`} - liczba [0, n],}
  \item{\code{`studiujący na studiach wieczorowych`} - liczba [0, n],}
  \item{\code{`studiujący na studiach zaocznych`} - liczba [0, n].}
}
}
\description{
Funkcja oblicza wskaźnik opisujący liczby badanych z grupy,
którzy kontynuowali naukę na studiach w poszczególnych trybach kształcenia.
}
\seealso{
\code{\link{studia_pierwsze}}
}
