% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{szkola_1_wyboru}
\alias{szkola_1_wyboru}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
szkola_1_wyboru(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja przechowująca rozkład odpowiedzi na pytanie W1 "Czy
obecna szkoła jest jedną z tych, do których najbardziej chciał(a) się Pan(i)
dostać?"
}
