% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{uczestnictwo_pnz}
\alias{uczestnictwo_pnz}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
uczestnictwo_pnz(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca liczbę absolwentów uczestniczących w danych
formach PNZ.
}
