% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{utworz_grupowanie_ze_zmiennej}
\alias{utworz_grupowanie_ze_zmiennej}
\title{Obliczanie wskaznikow na poziomie zagregowanym - definiowanie grupowania}
\usage{
utworz_grupowanie_ze_zmiennej(
  x,
  zmGrupujaca,
  zmGrupaOdniesienia,
  wykluczGrupeZGrupyOdniesienia = TRUE,
  ...
)
}
\arguments{
\item{x}{ramka danych (zwykle zbiór wskaźników indywidualnych)}

\item{zmGrupujaca}{nazwa zmiennej, której wartości definiują podział na
grupy, podana jako wyrażenie lub ciąg znaków}

\item{zmGrupaOdniesienia}{nazwa zmiennej, której wartości definiują grupę
odniesienia, podana jako wyrażenie lub ciąg znaków}

\item{wykluczGrupeZGrupyOdniesienia}{wartość logiczna - czy obserwacje
z analizowanej grupy powinny zostać wykluczone z grupy odniesienia}

\item{...}{opcjonalnie specyfikacja dodatkowych kolumn z ramki danych
podanej argumentem \code{x}, które mają zostać dołączone do zwracanej ramki
danych (podane w dowolny sposób akceptowany przez funkcję
\code{\link[dplyr]{select}})}
}
\value{
ramka danych, która może zostać użyta jako argument \code{grupy}
w wywołaniu funkcji \code{\link{agreguj_wskazniki}}.
}
\description{
Funkcja pozwala przygotować ramkę danych opisującą podział na
grupy, dla których mają następnie zostać obliczone wskaźniki zagregowane
przy pomocy funkcji \code{\link{agreguj_wskazniki}}, oraz odpowiadające im
grupy odniesienia w prostej, aczy typowej sytuacji, kiedy taki podział
definiują różne wartości jednej zmiennej w zbiorze (np. identyfikator szkoły),
a grupa odniesienia daje się określić przez taką samą (jak w analizowanej
grupie) wartość innej zmiennej (np. typ szkoły). Dodatkowy argument pozwala
określić czy obserwacje należące do analizowanej grupy powinny zostać,
wykluczone, czy też włączone do grupy odniesienia.
}
