% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_wyniki_pilrm.R
\name{wczytaj_wyniki_pilrm}
\alias{wczytaj_wyniki_pilrm}
\title{Wczytuje zbiór z wynikami pierwszej rundy monitorowania losow
absolwentow}
\usage{
wczytaj_wyniki_pilrm(x)
}
\arguments{
\item{x}{nazwa pliku z wynikami (ciąg znaków)}
}
\value{
lista ramek danych
}
\description{
Funkcja wczytuje plik z wynikami pilotażowej rundy monitoringu losów
absolwentóW i przekształca go na zbiory: czasów odpowiedzi, epizodów,
członków gospodarstw domowych i zbiór obejmujący pozostałe zmienne
}
