% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl.R
\name{wskaznik_bdl}
\alias{wskaznik_bdl}
\title{Pobieranie danych z Banku Danych Lokalnych GUS}
\usage{
wskaznik_bdl(id, zamienNazwe = NULL)
}
\arguments{
\item{id}{liczba całkowita - id wskaźnika w API BDL}

\item{zamienNazwe}{opcjonlanie ciąg znaków - nazwa, która ma zostać
przypisana wskaźnikowi zamiast tej, którą ma w API BDL}
}
\value{
ramka danych z przypisaną klasą "wskaznikiBDL"
}
\description{
Funkcja pozwala przy pomocy API BDL pobrać dane dot. wskaźnika, którego nie
da się zidentyfikować na podstawie samej jego nazwy (używając funkcji
\code{\link{znajdz_wskazniki_bdl}}).
}
\details{
Niektóre wskaźniki mają w BDL GUS przypisane niezwykle opisowe nazwy
w rodzaju "ogółem".  Wskaźniki takie (dokładie ich `id`) trzeba znaleźć
\emph{ręcznie} w API BDL, a następnie pobrać ich dane identyfikacyjne
korzysając właśnie z tej funkcji, aby można je wykorzystać w analogiczny
sposób, jak te, informacje o których pobiera się przy pomocy funkcji
\code{\link{znajdz_wskazniki_bdl}}.
}
\seealso{
\code{\link{pobierz_dane_bdl}}, \code{\link{znajdz_wskazniki_bdl}}
}
