% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{wskazniki_nie_z_epizodow}
\alias{wskazniki_nie_z_epizodow}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
wskazniki_nie_z_epizodow(x, maksRokEgz)
}
\arguments{
\item{x}{ramka danych z odpowiedziami na ankietę absolwentów - typowo element
\code{dane} listy zwracanej przez funkcję \code{\link{imputuj_miesiac_pk_1rm}}}

\item{maksRokEgz}{liczba - najpóźniejszy rok, w którym absolwent mógł zdać
egzamin, aby w wynikowym wskaźniku został oznaczony jako osoba, która ten
egzamin zdała}
}
\value{
data frame zawierająca następujące kolumny:
\itemize{
  \item{wszystkie kolumny ramki danych przekazanej argumentem \code{x},
        których nazwy \strong{nie} zaczynają się od "ABS_",}
  \item{\code{SZK_teryt} - nr TERYT powiatu, przeliczony na podstawie
        znajdującego się wcześniej w tej zmiennej nr TERYT gminy, na terenie
        której znajduje się szkoła,}
  \item{\code{UCZ_plec} - płeć ucznia ("M" lub "K"),}
  \item{\code{matura_zdana} - wskaźnik opisujący, czy badany zdał maturę
        (liczba: 1 - zdał, 0 - nie zdał),}
  \item{\code{egz_zaw_zdany} - wskaźnik opisujący, czy badany zdał egzamin
        zawodowy, tj. zdał wszystkie egzaminy niezbędne do uzyskania
        dyplomu potwierdzającego kwalifikacje w zawodzie (liczba: 1 - zdał,
        0 - nie zdał).}
}
}
\description{
Funkcja oblicza zmienne-wskaźniki, które wymagają informacji
nie ze zbioru epizodów.
}
\seealso{
\code{\link{egz_zaw_zdawalnosc}}, \code{\link{matura_zdawalnosc}},
\code{\link{liczba_kobiet}}
}
