% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{zawod_liczebnosc}
\alias{zawod_liczebnosc}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
zawod_liczebnosc(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca liczbę absolwentów danego zawodu.
}
