% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_2rm.R
\name{zawod_przygotowanie_szkola}
\alias{zawod_przygotowanie_szkola}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
zawod_przygotowanie_szkola(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym}
}
\value{
lista
}
\description{
Funkcja obliczająca odsetek absolwentów oceniających
przygotowanie do wykonywania zawodu jako "doskonałe" lub "dobre" oraz
podstawę procentowania. Wynik funkcji jest wsadem do tabeli 1 w raporcie.
}
