% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{zawody}
\alias{zawody}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
zawody(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}
}
\value{
ciąg znaków - nazwy zawodów, oddzielone przecinkami, ułożone
w kolejności od najczęściej do najrzadziej wystepującego
}
\description{
Funkcja oblicza wskaźnik opisujący wszystkie zawody, w których
uczyli się uczniowie w grupie.
}
