% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl.R
\name{znajdz_wskazniki_bdl}
\alias{znajdz_wskazniki_bdl}
\title{Pobieranie danych z Banku Danych Lokalnych GUS}
\usage{
znajdz_wskazniki_bdl(nazwa)
}
\arguments{
\item{nazwa}{ciąg znaków - fragment nazwy wskaźnika, który ma być znaleziony}
}
\value{
ramka danych z przypisaną klasą "wskaznikiBDL"
}
\description{
Korzystając z API BDL funkcja zwraca informacje o wskaźnikach, których
nazwa zawiera podany ciąg znaków.
}
\details{
\strong{Uwaga!} Nie wszystkie wskaźniki BDL GUS da się rozsądnie znaleźć
w ten sposób, gdyż niektóre mają przypisane niezwykle opisowe nazwy w rodzaju
"ogółem" - te trzeba by wcześniej znajdować poprzez nazwę grupy i podgrupy
wskaźników, jednak możliwości takiego przeszukiwania nie zaimplementowano
w pakiecie MLASZdane. Wskaźniki takie (dokładie ich `id`) trzeba znaleźć
\emph{ręcznie} w API BDL, a następnie pobrać ich dane identyfikacyjne
korzysając z funkcji \code{\link{wskaznik_bdl}}.
}
\seealso{
\code{\link{pobierz_dane_bdl}}, \code{\link{wskaznik_bdl}}
}
