% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{zwroc_wywolanie_grupy}
\alias{zwroc_wywolanie_grupy}
\title{Obliczanie wskaznikow na poziomie zagregowanym - funkcje pomocnicze}
\usage{
zwroc_wywolanie_grupy(x)
}
\arguments{
\item{x}{wektor (lub lista) jednoelementowy}
}
\value{
call
}
\description{
Nieeksportowana funkcja używana w ramach
\code{\link{agreguj_wskazniki}} i \code{\link{sprawdz_definicje_grup}} do
konwersji różnych form, w jakich mogą być zapisane wyrażenia definiujące
grupy na \emph{call}.
}
\seealso{
\code{\link{agreguj_wskazniki}},
\code{\link{sprawdz_definicje_grup}}
}
