% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{agreguj_wskazniki}
\alias{agreguj_wskazniki}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
agreguj_wskazniki(
  wskazniki,
  grupy,
  ...,
  wielowatkowo = FALSE,
  uzywajPakietow = vector(mode = "character", length = 0)
)
}
\arguments{
\item{wskazniki}{ramka danych ze wskaźnikami na poziomie idywidualnym
zwracana przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}}

\item{grupy}{ramka danych zawierająca definicje podziałów na grupy (oraz
ewentualnie inne zmienne, które zostaną dołączone do zwracanych zbiorów) -
zwrócona przez funkcję \code{\link{utworz_grupowanie_ze_zmiennej}} lub
przygotowana samodzielnie; musi zawierać kolumny \emph{grupa}
i \emph{odniesienie}, które zawierają wyrażenia wybierające jednostki
obserwacji należące do danej grupy lub grupy odniesienia (mogą być one
zapisane \emph{stricte} w formie niezewaluowanych wyrażeń języka, ale także
formuł lub ciągów znaków)}

\item{...}{wyrażenia postaci
\code{nazwa_wskaznika = funkcja_obliczajaca_wskaznik(.data, ew_inne_argumenty)}}

\item{wielowatkowo}{wartość logiczna - czy obliczenia powinny zostać wykonane
wielowątkowo}

\item{uzywajPakietow}{wektor ciągów znaków z nazwami pakietów, które musza
zostac załadowane, aby móc wykonać wyrażenia podane w {...} (z wyjątkiem
ładowanych domyślnie do sesji R oraz \emph{MLASZdane}, który zostanie
załadowany zawsze) - potrzebne tylko przy używaniu wielowątkowości
w systemach operacyjnych Windows}
}
\value{
lista dwóch ramek danych:
\itemize{
  \item{\code{grupy} - ramka danych zawierająca wskaźniki obliczone dla
        poszczególnych grup,}
  \item{\code{grupyOdniesienia} - ramka danych zawierająca wskaźniki
        obliczone dla odpowiadających im grup odniesienia.}
}
}
\description{
Funkcja oblicza wartości wskaźników na poziomie zagregowanym
na podstawie ramki danych ze wskaźnikami z poziomu indywidualnego
(zwrócnej przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}).
}
\seealso{
Funkcje, które wywołują \code{agreguj_wskazniki}:
\itemize{
  \item{\code{\link{agreguj_wskazniki_1rm}}.}
}
}
