% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki.R
\name{agreguj_wskazniki_1rm}
\alias{agreguj_wskazniki_1rm}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
agreguj_wskazniki_1rm(wskazniki, ...)
}
\arguments{
\item{wskazniki}{ramka danych ze wskaźnikami na poziomie idywidualnym
zwracana przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}}

\item{...}{zmienne definiujące podział na grupy, w ramach których ma zostać
przeprowadzona agregacja, występujące w ramce danych przekazanej argumentem
\code{wskazniki} - podane jako wyrażenia lub ciągi znaków}
}
\value{
data frame
}
\description{
Funkcja oblicza wartości wskaźników na poziomie zagregowanym
na podstawie ramki danych ze wskaźnikami z poziomu indywidualnego
(zwrócnej przez funkcję \code{\link{oblicz_wskazniki_ind_1rm}}).
}
\seealso{
Funkcje używane wewnętrznie do obliczania wskaźników zagregowanych:
\itemize{
  \item{\code{\link{liczba_zbadanych}},}
  \item{\code{\link{liczba_kobiet}},}
  \item{\code{\link{liczba_szkol}},}
  \item{\code{\link{zawody}},}
  \item{\code{\link{praca_nauka}},}
  \item{\code{\link{egz_zaw_zdawalnosc}},}
  \item{\code{\link{matura_zdawalnosc}},}
  \item{\code{\link{praca_przed_ukonczeniem_szkoly}},}
  \item{\code{\link{praca_czas_rozp}},}
  \item{\code{\link{praca_forma}},}
  \item{\code{\link{praca_forma2}},}
  \item{\code{\link{praca_zamieszkanie}},}
  \item{\code{\link{praca_zgodna_z_wyksztalceniem}},}
  \item{\code{\link{praca_zarobki}},}
  \item{\code{\link{praca_spelnienie_oczekiwan}},}
  \item{\code{\link{bezrobocie}},}
  \item{\code{\link{czas_agregacja}},}
  \item{\code{\link{czas_rozklad}},}
  \item{\code{\link{nauka}},}
  \item{\code{\link{studia_gdzie}},}
  \item{\code{\link{studia_odplatnosc}},}
  \item{\code{\link{studia_tryb}}.}
}
}
