% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputuj_mies_pk_pilrm.R
\name{imputuj_mies_pk_pilrm}
\alias{imputuj_mies_pk_pilrm}
\title{Imputacja miesiacow rozpoczecia i konca epizodow w danych sondazowych}
\usage{
imputuj_mies_pk_pilrm(x)
}
\arguments{
\item{x}{lista zwracana przez funkcję \code{\link{wczytaj_wyniki_pilrm}}}
}
\value{
ramka danych
}
\description{
Funkcja imputuje braki danych w zmiennych opisujących miesiąc rozpoczęcia
i miesiąc zakończenia epizodów pracy nauki i bezrobocia w danych sondażowych
z pilotażowej rundy monitringu.
}
