% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputuj_miesiac_pk_1rm.R
\name{imputuj_miesiac_pk_1rm}
\alias{imputuj_miesiac_pk_1rm}
\title{Imputacja miesiacow rozpoczecia i zakonczenia epizodow w danych sondazowych}
\usage{
imputuj_miesiac_pk_1rm(x, print = TRUE)
}
\arguments{
\item{x}{lista zwracana przez funkcję \code{\link{wczytaj_wyniki_1rm}}}

\item{print}{wartość logiczna - czy drukować na konsolę tabele z wynikami
imputacji? (warto wyłączyć przy uruchamianiu w ramach testów)}
}
\value{
lista ramek danych o takiej samej strukturze, jak zwracana przez
\code{\link{wczytaj_wyniki_1rm}}
}
\description{
Funkcja imputuje braki danych w zmiennych opisujących miesiąc rozpoczęcia
i miesiąc zakończenia epizodów pracy nauki i bezrobocia w danych sondażowych
z pilotażowej rundy monitringu.
}
\details{
Patrz dokumentacja dot. struktury i zawartości zbiorów danych
z \href{../doc/runda_1-dokumentacja.html}{1. rundy monitoringu}
tworzonych przy pomocy funkcji pakietu MLASZdane.
}
