% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany.R
\name{liczba_kobiet}
\alias{liczba_kobiet}
\title{Obliczanie wskaznikow na poziomie zagregowanym}
\usage{
liczba_kobiet(x)
}
\arguments{
\item{x}{ramka danych ze wskaźnikami na poziomie indywidualnym - typowo
grupa wyróżniona w ramach obiektu zwracanego przez funkcję
\code{\link{oblicz_wskazniki_ind_1rm}}}
}
\value{
liczba
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę zbadanych w grupie.
}
\seealso{
\code{\link{wskazniki_nie_z_epizodow}}
}
