% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polacz_dane.R
\name{polacz_dane}
\alias{polacz_dane}
\title{laczy dane ZUS, o wynikach egzaminow, uczniach, studentach i szkolach
oraz statystyki powiatow}
\usage{
polacz_dane(
  absolwenci,
  zus,
  zdajacyEgzaminy,
  uczniowie,
  studenci,
  szkoly,
  pnaPowiaty,
  dataMin,
  dataMax
)
}
\arguments{
\item{absolwenci}{dane opisujące absolwentów}

\item{zus}{dane o zatrudnieniu w formacie absolwento-miesiąco-płatników}

\item{zdajacyEgzaminy}{dane o podchodzących do egzaminów}

\item{uczniowie}{dane o uczniach (kontynuujących naukę)}

\item{studenci}{dane o studentach (kontynuujących naukę)}

\item{szkoly}{dane szkół}

\item{pnaPowiaty}{dane o powiatach}

\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch
znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch
znaków, np. '2015-09-30')}
}
\value{
data frame - połączone dane
}
\description{
Aby poprawnie przyłączyć statystyki powiatów macierz danych
uzupełniana jest tak, aby dla każdego absolwenta zawierała informacje o
każdym okresie w zadanym przedziale czasu, ale nie później niż do "końca
obserwacji w ZUS" (zakładając zamieszkanie absolwenta w uzupełnianych
okresach "w Polsce", kodowane jako pna = -1, oraz wszelkie cechy tytułu
ubezpieczenia ZUS równe 0, w szczególności także bezrob = 0).
}
