% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_indywidualny.R
\name{praca_forma_ind}
\alias{praca_forma_ind}
\title{Obliczanie wskaznikow na poziomie indywidualnym}
\usage{
praca_forma_ind(x, idAbsolwenta = "ID_RESP")
}
\arguments{
\item{x}{ramka danych z epizodami - typowo element \code{epizody} listy
zwracanej przez funkcję \code{\link{imputuj_miesiac_pk_1rm}}}

\item{idAbsolwenta}{nazwa zmiennej identyfikującej badanych (podana jako ciąg
znaków lub wyrażenie)}
}
\value{
data frame
}
\description{
Funkcja oblicza zmienne-wskaźniki opisujące, czy badany
kiedykolwiek był zatrudniony w określonej formie
}
\details{
W kodowaniu wykorzystywane są zarówno odpowiedzi na pytanie PG2H, jak i na
pytanie PG2G, z wykorzystaniem następujących reguł:
\itemize{
  \item{\code{pg2gh.1 = any(pg2h \%in\% 1)},}
  \item{\code{pg2gh.2 = any(pg2h \%in\% 2)},}
  \item{\code{pg2gh.3 = any(pg2h \%in\% 3)},}
  \item{\code{pg2gh.4 = any(pg2h \%in\% 4)},}
  \item{\code{pg2gh.5 = any(pg2h \%in\% 5)},}
  \item{\code{pg2gh.6 = any(pg2g \%in\% 6)},}
  \item{\code{pg2gh.7 = any(pg2g \%in\% 7)},}
  \item{\code{pg2gh.8 = any(pg2h \%in\% (7:8))},}
  \item{\code{pg2gh.9 = any(pg2h \%in\% 9 | is.na(pg2h))}.}
}
Dodatkowo, jeśli zmienna \code{pg2h} przyjmuje wartość 6 (praca bez umowy),
wszystkie wynikowe zmienne są dla danego respondenta przekodowywane na brak
danych.
}
\seealso{
\code{\link{praca_miesiac}}, \code{\link{praca_pierwsza}},
\code{\link{praca_ostatnia}}
}
