% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_zbior_osobo_miesiecy_1rm.R
\name{przygotuj_zbior_osobo_miesiecy_1rm}
\alias{przygotuj_zbior_osobo_miesiecy_1rm}
\title{Tworzenie zbioru osobo-miesiecy}
\usage{
przygotuj_zbior_osobo_miesiecy_1rm(x, print = TRUE)
}
\arguments{
\item{x}{lista zwracana przez funkcję \code{\link{imputuj_miesiac_pk_1rm}}}

\item{print}{wartość logiczna - czy drukować na konsolę tabele
z podsumowaniami przekształcanych danych? (warto wyłączyć przy uruchamianiu
w ramach testów)}
}
\value{
ramka danych
}
\description{
Funkcja przekształca dane o epizodach nauki, pracy i bezrobocia, zebrane
w ramach 1. rundy monitoringu na zbiór, w którym obserwacją jest konkretny
badany w konkretnym miesiącu, a zmienne opisują jego status zatrudnienia/nauki.
}
\details{
Patrz dokumentacja dot. struktury i zawartości zbiorów danych
z \href{../doc/runda_1-dokumentacja.html}{1. rundy monitoringu}
tworzonych przy pomocy funkcji pakietu MLASZdane.
}
