% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_zbior_osobo_miesiecy_pilrm.R
\name{przygotuj_zbior_osobo_miesiecy_pilrm}
\alias{przygotuj_zbior_osobo_miesiecy_pilrm}
\title{Tworzenie zbioru osobo-miesiecy}
\usage{
przygotuj_zbior_osobo_miesiecy_pilrm(x)
}
\arguments{
\item{x}{lista zwracana przez funkcję \code{\link{imputuj_miesiac_pk_pilrm}}}
}
\value{
ramka danych
}
\description{
Funkcja przekształca dane o epizodach nauki, pracy i bezrobocia, zebrane
w ramach pilotażowej rundy monitoringu na zbiór, w którym obserwacją jest konkretny
badany w konkretnym miesiącu, a zmienne opisują jego status zatrudnienia/nauki.
}
\details{
Patrz dokumentacja dot. struktury i zawartości zbiorów danych
z \href{../doc/runda_pilotazowa-dokumentacja.html}{rundy pilotażowej}
tworzonych przy pomocy funkcji pakietu MLASZdane.
}
