% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_wyniki_1rm.R
\name{wczytaj_wyniki_1rm}
\alias{wczytaj_wyniki_1rm}
\title{Wczytuje zbior z wynikami 1. rundy monitorowania losow absolwentow}
\usage{
wczytaj_wyniki_1rm(x)
}
\arguments{
\item{x}{nazwa pliku z wynikami (ciąg znaków)}
}
\value{
lista ramek danych
}
\description{
Funkcja wczytuje plik z wynikami 1. rundy monitoringu losów
absolwentów i przekształca go na zbiory: epizodów, członków gospodarstw
domowych i zbiór obejmujący pozostałe zmienne. \strong{Uwaga}, funkcja
zakłada, że \strong{do zbioru zostały już dopisane kody zawodów} (zakodowane
z odpowiedzi otwartych udzielonych przez respondentów).
}
\details{
Funkcja jest całkowicie nieodporna na niewystępowanie w zbiorze kolumn,
których istnienia oczekuje - w takiej sytuacji będzie rzucać umiarkowanie
informatywnymi błędami. Nazwa ostatecznej wersji zbioru z wynikami badania,
do wykorzystania, na której testowane było działanie funkcji to:
\emph{MLEZAiMD_I_runda_CAPI_absolwent_n7713_20180924_z_wagami_z_kodowaniem.sav}.

Więcej szczegółów - patrz dokumentacja dot. struktury i zawartości zbiorów
danych z \href{../doc/runda_1-dokumentacja.html}{1. rundy monitoringu}
tworzonych przy pomocy funkcji pakietu MLASZdane.
}
